# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AIGYPT Education Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  code: string;
}

export interface LoginResult {
  success: boolean;
  memberType: string;
  /** @nullable */
  batchNumber?: number | null;
  /** @nullable */
  name?: string | null;
}

export interface Member {
  email: string;
  /** @nullable */
  name?: string | null;
  memberType: string;
  /** @nullable */
  batchNumber?: number | null;
}

export interface Progress {
  sesiNumber: number;
  isCompleted: boolean;
  /** @nullable */
  completedAt?: string | null;
}

export interface ProgressInput {
  sesiNumber: number;
}

export type GenerateCodesInputType = typeof GenerateCodesInputType[keyof typeof GenerateCodesInputType];


export const GenerateCodesInputType = {
  mandiri: 'mandiri',
  kelas: 'kelas',
} as const;

export interface GenerateCodesInput {
  type: GenerateCodesInputType;
  /** @nullable */
  batchNumber?: number | null;
  /**
     * @minimum 1
     * @maximum 100
     */
  count: number;
}

export interface GenerateCodesResult {
  codes: string[];
}

export interface AccessCode {
  id: string;
  code: string;
  type: string;
  /** @nullable */
  batchNumber?: number | null;
  isUsed: boolean;
  /** @nullable */
  usedByEmail?: string | null;
  /** @nullable */
  usedAt?: string | null;
  createdAt: string;
  /** @nullable */
  expiresAt?: string | null;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}