# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AIGYPT Education Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: progress
    description: Learning progress tracking
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /progress:
    get:
      operationId: getProgress
      tags: [progress]
      summary: Get learning progress for current user
      responses:
        "200":
          description: Progress list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Progress"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: markComplete
      tags: [progress]
      summary: Mark a session as complete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressInput"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/generate-codes:
    post:
      operationId: generateCodes
      tags: [admin]
      summary: Generate access codes (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCodesInput"
      responses:
        "200":
          description: Generated codes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateCodesResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/codes:
    get:
      operationId: listCodes
      tags: [admin]
      summary: List all access codes (admin only)
      responses:
        "200":
          description: Access codes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessCode"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginInput:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string

    LoginResult:
      type: object
      required: [success, memberType]
      properties:
        success:
          type: boolean
        memberType:
          type: string
        batchNumber:
          type: ["integer", "null"]
        name:
          type: ["string", "null"]

    Member:
      type: object
      required: [email, memberType]
      properties:
        email:
          type: string
        name:
          type: ["string", "null"]
        memberType:
          type: string
        batchNumber:
          type: ["integer", "null"]

    Progress:
      type: object
      required: [sesiNumber, isCompleted]
      properties:
        sesiNumber:
          type: integer
        isCompleted:
          type: boolean
        completedAt:
          type: ["string", "null"]

    ProgressInput:
      type: object
      required: [sesiNumber]
      properties:
        sesiNumber:
          type: integer

    GenerateCodesInput:
      type: object
      required: [type, count]
      properties:
        type:
          type: string
          enum: [mandiri, kelas]
        batchNumber:
          type: ["integer", "null"]
        count:
          type: integer
          minimum: 1
          maximum: 100

    GenerateCodesResult:
      type: object
      required: [codes]
      properties:
        codes:
          type: array
          items:
            type: string

    AccessCode:
      type: object
      required: [id, code, type, isUsed, createdAt]
      properties:
        id:
          type: string
        code:
          type: string
        type:
          type: string
        batchNumber:
          type: ["integer", "null"]
        isUsed:
          type: boolean
        usedByEmail:
          type: ["string", "null"]
        usedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        expiresAt:
          type: ["string", "null"]

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string